+++ modules/snooker/src/main/ts/ephox/snooker/api/OtherCells.ts
@@ -1,19 +1,19 @@
 import { Arr, Optional } from '@ephox/katamari';
 import { SugarElement } from '@ephox/sugar';
 
+import * as GridRow from '../model/GridRow';
 import { onCells, TargetSelection, toDetailList } from '../model/RunOperation';
 import * as Transitions from '../model/Transitions';
-import { CellElement } from '../util/TableTypes';
 import { Generators } from './Generators';
 import { DetailExt, RowCells } from './Structs';
 import { Warehouse } from './Warehouse';
 
 export interface OtherCells {
-  readonly upOrLeftCells: SugarElement<CellElement>[];
-  readonly downOrRightCells: SugarElement<CellElement>[];
+  readonly upOrLeftCells: SugarElement<HTMLTableCellElement>[];
+  readonly downOrRightCells: SugarElement<HTMLTableCellElement>[];
 }
 
-const getUpOrLeftCells = (grid: RowCells[], selectedCells: DetailExt[]): SugarElement<CellElement>[] => {
+const getUpOrLeftCells = (grid: RowCells<HTMLTableRowElement>[], selectedCells: DetailExt[]): SugarElement<HTMLTableCellElement>[] => {
   // Get rows up or at the row of the bottom right cell
   const upGrid = grid.slice(0, selectedCells[selectedCells.length - 1].row + 1);
   const upDetails = toDetailList(upGrid);
@@ -24,7 +24,7 @@ const getUpOrLeftCells = (grid: RowCells[], selectedCells: DetailExt[]): SugarEl
   });
 };
 
-const getDownOrRightCells = (grid: RowCells[], selectedCells: DetailExt[]): SugarElement<CellElement>[] => {
+const getDownOrRightCells = (grid: RowCells<HTMLTableRowElement>[], selectedCells: DetailExt[]): SugarElement<HTMLTableCellElement>[] => {
   // Get rows down or at the row of the top left cell (including rowspans)
   const downGrid = grid.slice(selectedCells[0].row + selectedCells[0].rowspan - 1, grid.length);
   const downDetails = toDetailList(downGrid);
@@ -41,8 +41,9 @@ const getOtherCells = (table: SugarElement<HTMLTableElement>, target: TargetSele
 
   return details.map((selectedCells) => {
     const grid = Transitions.toGrid(warehouse, generators, false);
-    const upOrLeftCells = getUpOrLeftCells(grid, selectedCells);
-    const downOrRightCells = getDownOrRightCells(grid, selectedCells);
+    const { rows } = GridRow.extractGridDetails(grid);
+    const upOrLeftCells = getUpOrLeftCells(rows, selectedCells);
+    const downOrRightCells = getDownOrRightCells(rows, selectedCells);
     return {
       upOrLeftCells,
       downOrRightCells