+++ modules/tinymce/src/core/main/ts/api/AddOnManager.ts
@@ -101,7 +101,7 @@ export interface UrlObject { prefix: string; resource: string; suffix: string }
 
 type WaitState = 'added' | 'loaded';
 
-export type AddOnConstructor<T> = (editor: Editor, url: string) => void | T;
+export type AddOnConstructor<T> = (editor: Editor, url: string) => T;
 
 interface AddOnManager<T> {
   items: AddOnConstructor<T>[];
@@ -302,5 +302,6 @@ AddOnManager.languageLoad = true;
 AddOnManager.baseURL = '';
 AddOnManager.PluginManager = AddOnManager<any>();
 AddOnManager.ThemeManager = AddOnManager<any>();
+AddOnManager.ModelManager = AddOnManager<any>();
 
 export default AddOnManager;