+++ modules/tinymce/src/core/main/ts/api/PublicApi.ts
@@ -40,6 +40,7 @@ import Styles from './html/Styles';
 import Writer, { WriterSettings } from './html/Writer';
 import IconManager from './IconManager';
 import { EditorModeApi } from './Mode';
+import ModelManager, { Model } from './ModelManager';
 import NotificationManager, { NotificationApi, NotificationSpec } from './NotificationManager';
 import { EditorOptions, RawEditorOptions } from './OptionTypes';
 import PluginManager, { Plugin } from './PluginManager';
@@ -115,6 +116,7 @@ export {
   Shortcuts,
   PluginManager,
   ThemeManager,
+  ModelManager,
   UndoManager,
   WindowManager,
   FakeClipboard,
@@ -135,5 +137,6 @@ export {
   Plugin,
   SchemaSettings,
   Theme,
+  Model,
   WriterSettings
 };