+++ modules/tinymce/src/plugins/table/test/ts/browser/ui/TableCellDialogTest.ts
@@ -4,8 +4,8 @@ import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
+import { TableEventData, TableModifiedEvent } from 'tinymce/core/api/EventTypes';
 import { EditorEvent } from 'tinymce/core/api/util/EventDispatcher';
-import { TableEventData, TableModifiedEvent } from 'tinymce/plugins/table/api/Events';
 import Plugin from 'tinymce/plugins/table/Plugin';
 
 import * as TableTestUtils from '../../module/test/TableTestUtils';
@@ -20,11 +20,11 @@ describe('browser.tinymce.plugins.table.TableCellDialogTest', () => {
       '*': 'width,height,vertical-align,text-align,float,border-color,border-style,background-color,border,padding,border-spacing,border-collapse,border-width'
     },
     setup: (editor: Editor) => {
-      editor.on('tablemodified', (event) => {
+      editor.on('TableModified', (event) => {
         logEventTypes(event);
         logTableModified(event);
       });
-      editor.on('newcell', logEventTypes);
+      editor.on('NewCell', logEventTypes);
     }
   }, [ Plugin ], true);